mm: Add a bitmap into mmu_notifier_{clear,test}_young
The bitmap is provided for secondary MMUs to use if they support it. For
test_young, it represents the pages that were young in the interval
[start, end). For clear_young, it represents the pages that we wish the
secondary MMU to clear the accessed/young bit for.

This allows MGLRU's look-around logic to work faster, resulting in a 4%
improvement in real workloads[1]. Also introduce MMU_NOTIFIER_YOUNG_FAST
to indicate to main mm that doing look-around logic is likely to be
beneficial.

If the secondary MMU doesn't support the bitmap, it must return
an int that contains MMU_NOTIFIER_YOUNG_BITMAP_UNRELIABLE.

[1]: https://lore.kernel.org/all/20230609005935.42390-1-yuzhao@google.com/

Co-developed-by: Yu Zhao <yuzhao@google.com>
Signed-off-by: James Houghton <jthoughton@google.com>