KVM: Add basic bitmap support into kvm_mmu_notifier_test/clear_young
kvm_arch_has_fast_test_age_gfn indicates that the architecture:
1. Supports walking the KVM page tables without taking the MMU lock.
2. Supports bitmap-based aging in kvm_mmu_notifier_test_clear_young.

Architectures need to opt in to this. If the architecture has not opted
in or is unable to handle lockless walks at that moment:
 1. If a bitmap was provided, we inform the caller that the bitmap is
    unreliable (MMU_NOTIFIER_YOUNG_BITMAP_UNRELIABLE).
 2. If a bitmap was not provided, fall back to the old logic.

Also add logic for architectures to easily use the bitmap provided, if
they are able. The expectation is that the architecture's implementation
of kvm_gfn_test_age will use kvm_gfn_record_young, and kvm_gfn_age will
use kvm_gfn_should_age.

Co-developed-by: Yu Zhao <yuzhao@google.com>
Signed-off-by: James Houghton <jthoughton@google.com>